requirements.txt
Django==3.1.2
django-cors-headers
numpy
scipy
Pillow